Add buf-size#boundedBlockLength as a provided feature.
This is to not block plugins that "require" `buf-size#boundedBlockLength`
as a feature.

It is already part of "options", but those options are never exposed as
provided features when it is instantiating a plugin, nor permitted as
a feature that a plugin can require.

sfizz and some Carla plugins are known to be refused as a result. This
change saves those plugins (not meaning that they then work; sfizz still
needs some way to pass filename, which is not clear to me).